Fix some compiler complaints, try to make C++ code 64-bit clean.

+++ src/IcuEC/IcuRegexEC.cpp
@@ -288,7 +288,7 @@ namespace IcuRegexEC
                 fprintf(stderr, "actual nLen %d\n", nLen);
                 rnOutLen = nLen;
                 //lpOutBuffer[rnOutLen] = '\0';   // null terminate in case extract didn't do it
-                fprintf(stderr, "lpOutBuffer length = %d\n", strlen(lpOutBuffer));
+                fprintf(stderr, "lpOutBuffer length = %u\n", (unsigned)strlen(lpOutBuffer));
                 fprintf(stderr, "lpOutBuffer: '%s'\n", lpOutBuffer);
             }
         }
